+++ src/helpers.rs
@@ -1,5 +1,5 @@
-///! Public helper utilities
-///!
+//! Public helper utilities
+//!
 use crate::models::{Cell, Column, Row};
 use crate::types::Result;
 
@@ -9,6 +9,7 @@ use std::io::{Error, ErrorKind};
 /// Define type aliases for the column mappings so that we can DRY.
 type ColumnNameToId<'a> = HashMap<&'a str, u64>;
 type ColumnIdToName<'a> = HashMap<u64, &'a str>;
+type ColumnNameToCell<'a> = HashMap<&'a str, &'a Cell>;
 
 /// **Column Mapper** - Utility to generate the `name` <-> `id` mappings for
 /// columns in a sheet.
@@ -65,6 +66,7 @@ impl<'a> ColumnMapper<'a> {
 ///
 pub struct CellGetter<'a> {
     column_name_to_id: &'a ColumnNameToId<'a>,
+    id_to_column_name: &'a ColumnIdToName<'a>,
 }
 
 impl<'a> CellGetter<'a> {
@@ -72,6 +74,7 @@ impl<'a> CellGetter<'a> {
     pub fn new(columns: &'a ColumnMapper<'a>) -> Self {
         Self {
             column_name_to_id: &columns.name_to_id,
+            id_to_column_name: &columns.id_to_name,
         }
     }
 
@@ -82,9 +85,20 @@ impl<'a> CellGetter<'a> {
 
     /// Create a new `CellGetter` from a reference to a mapping of *column name*
     /// to *column id*
-    pub fn from_name_to_id(column_name_to_id: &'a ColumnNameToId<'a>) -> Self {
-        Self { column_name_to_id }
-    }
+    ///
+    /// NOTE: Disabling this for now, because I can't get the lifetimes to work.
+    // pub fn from_name_to_id<'b>(column_name_to_id: &'b ColumnNameToId<'b>) -> CellGetter<'a> {
+    //     let mut id_to_column_name: ColumnIdToName<'b> =
+    //         HashMap::with_capacity(column_name_to_id.len());
+    //     for (&name, &id) in column_name_to_id {
+    //         id_to_column_name.insert(id, name);
+    //     }
+    //
+    //     Self {
+    //         column_name_to_id,
+    //         id_to_column_name: &id_to_column_name,
+    //     }
+    // }
 
     /// Find a `cell` in a `row` by its *column name*.
     pub fn by_name<'b>(&'a self, row: &'a Row, name: &'b str) -> Result<&Cell> {
@@ -100,5 +114,21 @@ impl<'a> CellGetter<'a> {
     /// Find a `cell` in a `row` by its *column id*.
     pub fn by_id(&'a self, row: &'a Row, column_id: u64) -> Result<&Cell> {
         row.get_cell_by_id(column_id)
+    }
+
+    /// Retrieve a mapping of *column name* to `Cell` object.
+    ///
+    /// Note: this is likely more efficient when multiple `Cell`s are to be
+    /// retrieved from an individual `Row`.
+    pub fn name_to_cell(&'a self, row: &'a Row) -> ColumnNameToCell<'a> {
+        let mut col_name_to_cell: ColumnNameToCell<'a> = HashMap::with_capacity(row.cells.len());
+
+        for cell in &row.cells {
+            if let Some(&col_name) = self.id_to_column_name.get(&cell.column_id) {
+                col_name_to_cell.insert(col_name, cell);
+            }
+        }
+
+        col_name_to_cell
     }
 }