+++ src/lib.rs
@@ -39,18 +39,61 @@
 //! - [List Columns](https://smartsheet-platform.github.io/api-docs/#list-columns)
 //! - [Get Sheet](https://smartsheet-platform.github.io/api-docs/#get-sheet)
 //! - [Get Row](https://smartsheet-platform.github.io/api-docs/#get-row)
+//! - [Get Column](https://smartsheet-platform.github.io/api-docs/#get-column)
 //!
 //! You can check out sample usage of these API methods in the [examples/](https://github.com/rnag/smartsheet-rs/tree/main/examples)
 //! folder in the project repo on GitHub.
 //!
-//! ## Dependencies
+//! ## A Larger Example
+//!
+//! This section contains more examples of usage. You can find it in the readme documentation on the
+//! [crates.io] page, or alternatively in the [`README.md`] file on the GitHub project repo.
+//!
+//! [crates.io]: https://crates.io/crates/smartsheet-rs#a-larger-example
+//! [`README.md`]: https://github.com/rnag/smartsheet-rs#a-larger-example
+//!
+//! ## Dependencies and Features
 //!
 //! This library uses only the minimum required dependencies, in order
 //! to keep the overall size small. This crate uses [hyper][] and
-//! [hyper-tls][] internally, to make HTTPS requests to the Smartsheet API.
+//! [hyper-rustls][] internally, to make HTTPS requests to the Smartsheet API.
+//!
+//! While `hyper-rustls` was chosen as the default TLS implementation
+//! because it works without issue when cross-compiling for the
+//! **x86_64-unknown-linux-musl** target as is common for [AWS Lambda][]
+//! deployments, it is still possible to instead use the native [`hyper-tls`][]
+//! implementation, which relies on OpenSSL.
+//!
+//! To do this, disable the default "rust-tls" feature and enable the "native-tls" feature:
+//!
+//! ```toml
+//! [dependencies]
+//! smartsheet-rs = { version = "0.3", default-features = false, features = ["native-tls", "logging", "serde-std"] }
+//! ```
 //!
 //! [hyper]: https://docs.rs/hyper
-//! [hyper-tls]: https://docs.rs/hyper-tls
+//! [hyper-rustls]: https://docs.rs/hyper-rustls
+//! [`hyper-tls`]: https://docs.rs/hyper-tls
+//! [AWS Lambda]: https://docs.aws.amazon.com/sdk-for-rust/latest/dg/lambda.html
+//!
+
+// #![warn(missing_docs)]
+
+mod features_check;
+
+#[cfg(feature = "logging")]
+mod log {
+    pub use log::{debug, error, trace, warn};
+}
+
+#[cfg(not(feature = "logging"))]
+mod log {
+    macro_rules! debug      ( ($($tt:tt)*) => {{}} );
+    macro_rules! error      ( ($($tt:tt)*) => {{}} );
+    macro_rules! trace      ( ($($tt:tt)*) => {{}} );
+    macro_rules! warning    ( ($($tt:tt)*) => {{}} );
+    pub(crate) use {debug, error, trace, warning as warn};
+}
 
 pub use api::SmartsheetApi;
 pub use helpers::{CellGetter, ColumnMapper};
@@ -60,6 +103,7 @@ pub mod auth;
 pub mod builders;
 pub mod constants;
 pub mod helpers;
+mod https;
 pub mod models;
 pub mod status;
 pub mod types;